Sub-Directories included for Platform and Network Configuraiton
This commit includes all the sub-directories that will be included in the platform and network configuraiton tab. Within each sub-directory will be markdown files based on those listed in NewtorkConfiguraiton.md and PlatformConfiguraiton.md. Currently no sub-directories were established for Platform configuraitons due to the limited number. But may be subject to changes.

Signed-off-by: Korel <Korelucpinar@gmail.com>