Explanation for Including NAT in VRRP Example
The example that had been included was attempting to show how a device could access the internet and maintain connection using VRRP. NAT was used to translate requests from a private ip address to a public ip address. An explanation was provided to clarify why this was needed. An explanaiton was also included to specify NAT is not required if you are only working within a LAN.

Signed-off-by: Korel <Korelucpinar@gmail.com>