Using DENT in GNS3 with a Remote Server
Adds the last page for using DENT in GNS3, this time with a Remote Server. This document contains a walkthrough for adding a remote GNS3 server. When using a remote server, it is also noted and recommended to use a VPN if accessing the internet. Additionally, a similar example is included as that shown using the GNS3 VM documentation. This example is three DENT appliances connected to one another and running on the Remote Server.

Signed-off-by: Korel <Korelucpinar@gmail.com>