IEEE-802.1x Port Based and MAC ADDRESS Authentication
This document covers how to estbalish an IEEE 802.1x standard for network security using Bridge Port Configurations and a Radius Server. A document is also included for MAC Address authentication which builds on the established work in the IEEE 802.1x document. Example configuraitons are given for both and outputs are tested and shown to explain how the desired configuraiton has been established.

Signed-off-by: Korel <Korelucpinar@gmail.com>