Addresses Taskin's Concerns regarding the First commit
This commit addresses the following concerns by Taskin regarding the first commit of the IEEE 802.1x documentation. It includes: 1. An explanation of how the Protocol works. 2. Clarifies info on where all the SW users will be installing are going to be at with a graphical representation. 3. Explains how each installed utility will be used. 4. Clarifies if a configuration is meant for the Supplicant, Authenticator, or Authentication-Server. 5. Reworded the iproute2 Installation and Appendix section to provide install information for the specific iproute2-6.9.0 version used during the configuration of the demo.

Signed-off-by: Korel <Korelucpinar@gmail.com>